+++ src/Kiota.Builder/CodeDOM/CodeUnionType.cs
@@ -1,24 +1,14 @@
 using System;
-using System.Collections.Generic;
-using System.Linq;
 
-namespace Kiota.Builder
-{
-    public class CodeUnionType : CodeTypeBase, ICloneable {
-        public void AddType(params CodeType[] codeTypes) {
-            EnsureElementsAreChildren(codeTypes);
-            foreach(var codeType in codeTypes.Where(x => x != null && !Types.Contains(x)))
-                types.Add(codeType);
-        }
-        private readonly List<CodeType> types = new ();
-        public IEnumerable<CodeType> Types { get => types; }
+namespace Kiota.Builder;
 
-        public override object Clone() {
-            var value = new CodeUnionType{
-            }.BaseClone<CodeUnionType>(this);
-            if(Types?.Any() ?? false)
-                value.AddType(Types.ToArray());
-            return value;
-        }
+/// <summary>
+/// The base class for union types. (anyOf multiple properties at a time)
+/// </summary>
+public class CodeUnionType : CodeComposedTypeBase, ICloneable {
+    public override object Clone() {
+        var value = new CodeUnionType{
+        }.BaseClone<CodeUnionType>(this);
+        return value;
     }
 }