+++ src/Kiota.Builder/Writers/Java/JavaConventionService.cs
@@ -34,7 +34,7 @@ public class JavaConventionService : CommonLanguageConventionService
 
     public override string GetTypeString(CodeTypeBase code, CodeElement targetElement, bool includeCollectionInformation = true)
     {
-        if(code is CodeUnionType) 
+        if(code is CodeComposedTypeBase) 
             throw new InvalidOperationException($"Java does not support union types, the union type {code.Name} should have been filtered out by the refiner");
         else if (code is CodeType currentType) {
             var typeName = TranslateType(currentType);